fix minor bugs
