+++ federatedscope/core/fed_runner.py
@@ -8,8 +8,7 @@ import numpy as np
 from federatedscope.core.workers import Server, Client
 from federatedscope.core.gpu_manager import GPUManager
 from federatedscope.core.auxiliaries.model_builder import get_model
-from federatedscope.core.auxiliaries.data_builder import merge_data
-from federatedscope.core.auxiliaries.utils import get_resource_info
+from federatedscope.core.auxiliaries.utils import get_resource_info, merge_data
 
 logger = logging.getLogger(__name__)
